requirements.txt
pyglet
PyOpenGL
pyshaders
pyglbuffers
pillow
hurry.filesize > 0.9